+++ packages/video_player/video_player_platform_interface/CHANGELOG.md
@@ -1,4 +1,4 @@
-## 6.0.3
+## 6.1.0
 
 * Aligns Dart and Flutter SDK constraints.
 * Adds the `VideoEventType.isPlayingStateUpdate` event to track changes in play / pause state with